requirements.txt
rank-bm25>=0.2.2
pyyaml>=6.0
transformers>=4.40.0
torch>=2.0.0
sentencepiece
protobuf
